+++ outputs.tf
@@ -3,19 +3,30 @@ output "mysql_administrator_login" {
   description = "Administrator login for MySQL server"
 }
 
+output "mysql_administrator_password" {
+  value       = local.administrator_password
+  description = "Administrator password for mysql server"
+  sensitive   = true
+}
+
+output "mysql_databases" {
+  value       = azurerm_mysql_database.mysql_db
+  description = "Map of databases infos"
+}
+
 output "mysql_databases_names" {
-  value       = azurerm_mysql_database.mysql_db[*].name
+  value       = [for db in azurerm_mysql_database.mysql_db : db.name]
   description = "List of databases names"
 }
 
 output "mysql_database_ids" {
   description = "The list of all database resource ids"
-  value       = azurerm_mysql_database.mysql_db[*].id
+  value       = [for db in azurerm_mysql_database.mysql_db : db.id]
 }
 
 output "mysql_firewall_rule_ids" {
-  value       = azurerm_mysql_firewall_rule.firewall_rules[*].id
-  description = "List of MySQL created rules"
+  value       = azurerm_mysql_firewall_rule.firewall_rules
+  description = "Map of MySQL created rules"
 }
 
 output "mysql_fqdn" {
@@ -33,22 +44,23 @@ output "mysql_server_name" {
   description = "MySQL server name"
 }
 
-output "mysql_vnet_rule_ids" {
-  value       = azurerm_mysql_virtual_network_rule.vnet_rules[*].id
-  description = "The list of all vnet rule resource ids"
+output "mysql_vnet_rules" {
+  value       = azurerm_mysql_virtual_network_rule.vnet_rules
+  description = "The map of all vnet rules"
 }
+
 output "mysql_databases_users" {
-  description = "List of usernames of created users corresponding to input databases names."
-  value       = local.db_users_login
+  description = "Map of user name for each database"
+  value       = { for db, c in var.databases : db => mysql_user.users[db].user }
 }
 
-output "mysql_databases_users_passwords" {
-  description = "List of passwords of created users corresponding to input databases names."
-  value       = random_password.db_passwords[*].result
+output "mysql_databases_logins" {
+  description = "Map of user login for each database"
+  value       = { for db, c in var.databases : db => format("%s@%s", mysql_user.users[db].user, azurerm_mysql_server.mysql_server.name) }
+}
+
+output "mysql_databases_passwords" {
+  description = "Map of user password for each database"
+  value       = { for db, c in var.databases : db => random_password.db_passwords[db].result }
   sensitive   = true
-}
-
-output "mysql_configuration_id" {
-  value       = azurerm_mysql_configuration.mysql_config[*].id
-  description = "The list of all configurations resource ids"
 }