+++ provider.tf
@@ -1,9 +1,9 @@
 provider "mysql" {
-  alias = "create-users"
+  alias = "users_mgmt"
 
   endpoint = format("%s:3306", azurerm_mysql_server.mysql_server.fqdn)
   username = local.administrator_login
-  password = var.administrator_password
+  password = local.administrator_password
 
   tls = var.force_ssl
 }