+++ r-dbusers.tf
@@ -1,31 +1,31 @@
 resource "random_password" "db_passwords" {
-  count = var.create_databases_users ? length(var.databases_names) : 0
+  for_each = toset(var.create_databases_users ? keys(var.databases) : [])
 
   special = false
   length  = 32
 }
 
 resource "mysql_user" "users" {
-  count = var.create_databases_users ? length(var.databases_names) : 0
+  for_each = toset(var.create_databases_users ? keys(var.databases) : [])
 
-  provider = mysql.create-users
+  provider = mysql.users_mgmt
 
-  user               = (var.enable_user_suffix ? format("%s_user", var.databases_names[count.index]) : var.databases_names[count.index])
-  plaintext_password = random_password.db_passwords[count.index].result
+  user               = format("%s%s", each.key, local.user_suffix)
+  plaintext_password = random_password.db_passwords[each.key].result
   host               = "%"
 
   depends_on = [azurerm_mysql_database.mysql_db, azurerm_mysql_firewall_rule.firewall_rules]
 }
 
 resource "mysql_grant" "roles" {
-  count = var.create_databases_users ? length(var.databases_names) : 0
+  for_each = toset(var.create_databases_users ? keys(var.databases) : [])
 
-  provider = mysql.create-users
+  provider = mysql.users_mgmt
 
-  user       = (var.enable_user_suffix ? format("%s_user", var.databases_names[count.index]) : var.databases_names[count.index])
+  user       = format("%s%s", each.key, local.user_suffix)
   host       = "%"
-  database   = var.databases_names[count.index]
-  privileges = ["ALL"]
+  database   = each.key
+  privileges = ["ALL PRIVILEGES"]
 
   depends_on = [mysql_user.users]
 }