+++ r-vnetrules.tf
@@ -1,12 +1,8 @@
 resource "azurerm_mysql_virtual_network_rule" "vnet_rules" {
-  count = length(var.allowed_subnets)
+  for_each = var.allowed_subnets
 
-  name = format(
-    "%s-%s",
-    element(split("/", var.allowed_subnets[count.index]), 8), # VNet name
-    element(split("/", var.allowed_subnets[count.index]), 10) # Subnet name
-  )
+  name                = each.key
   resource_group_name = var.resource_group_name
   server_name         = azurerm_mysql_server.mysql_server.name
-  subnet_id           = var.allowed_subnets[count.index]
+  subnet_id           = each.value
 }